+++ Components/IRacing/Plugins/GameState/Session.cs
@@ -6,5 +6,16 @@
         public bool TimeLimited { get; set; }
         public double TotalSessionTime { get; set; }
         public int TotalSessionLaps { get; set; }
+
+        public Session Clone()
+        {
+            return new Session
+            {
+                LapsLimited = LapsLimited,
+                TimeLimited = TimeLimited,
+                TotalSessionTime = TotalSessionTime,
+                TotalSessionLaps = TotalSessionLaps,
+            };
+        }
     }
 }