+++ CHANGELOG.md
@@ -26,6 +26,11 @@
   - New plugin: DiscordPlugin
   - IRacingPlugin: Fix IRacingWeatherInfo - now reports actual values and not values from start of session.
   - IRacingPlugin: Support sending "raw" game state events. Allowing you to act on data changes not directly represented as an event.
+  - IRacingPlugin: Rename IRacingCarCompletedLap to IRacingCompletedLap
+  - IRacingPlugin: Renamed FuelDiff to FuelDelta in IRacingCompletedLap, IRacingPitstopReport for consistency
+  - IRacingPlugin: Adds BestLap to IRacingCompletedLap.
+  - IRacingPlugin: Rename IRacingCompletedLap.Time to IRacingCompletedLap.LapTime.
+  - IRacingPlugin: Add IRacingCompletedLap.EstimatedLapTime - to show if this was a laptime as recorded by IRacing or by Slipstream in case IRacing doesn't provide a laptime
 
 ## [0.4.0](https://github.com/dennis/slipstream/releases/tag/v0.3.0) (2020-01-10)
 [Full Changelog](https://github.com/dennis/slipstream/compare/v0.3.0...v0.4.0)